refactor(core): split TUI types and secret redaction out of core
core/types.ts mixed domain records with TUI routing types, so ScreenName
and RouterContextValue leaked into the public API and into the Windows
package's synced shared/types.ts, a package with no TUI. They now live
in tui/types.ts; index.ts re-exports them from the new home so the
public surface is unchanged, and the win copy drops 55 lines of TUI
types after sync:shared.

redactSecrets moved from utils/display.ts to core/redact.ts: it is a
persistence security invariant, not a display concern, and the old home
made the JSON store depend transitively on the Ink color palette. The
win drift test now reads core/redact.ts for its verbatim checks; its
redaction tests moved to test/redact.test.ts alongside.

No behavior change; test counts identical (741 root, 51 win).
