+++ test/router.test.ts
@@ -6,7 +6,7 @@ import { join } from 'node:path'
 import { tmpdir } from 'node:os'
 import { nextHistoryOnPush, nextHistoryOnReset } from '../src/tui/router.js'
 import { Router } from '../src/tui/router.js'
-import type { ScreenName } from '../src/core/types.js'
+import type { ScreenName } from '../src/tui/types.js'
 
 function state(entries: ScreenName[], index: number) {
   return { entries, index }
