+++ test/screens/Agent-Output.test.tsx
@@ -8,7 +8,8 @@ import { AgentOutput } from '../../src/tui/screens/run/AgentOutput.js'
 import { ToastProvider } from '../../src/tui/contexts/ToastContext.js'
 import * as runsState from '../../src/core/runs.js'
 import * as runtime from '../../src/core/runtime.js'
-import type { RunRecord, AgentRecord, RouterContextValue } from '../../src/core/types.js'
+import type { RunRecord, AgentRecord } from '../../src/core/types.js'
+import type { RouterContextValue } from '../../src/tui/types.js'
 
 const waitForInput = () => vi.advanceTimersByTimeAsync(75)
